Ex 3.5 free used memory on module exit
As suggested by Franciaso Giana the `fifo` should not keep it's internal
data allocated after unloading it. This commit is an attempt to solve
this situation.

More reading on the internal list and allocation parts of the kernel is needed
to succesfully affirm that this is a working solution, but it seemed
like the obvious code to write.